Refactor ConnectDapp.jsx
